Get id from token for creating and auto-saving applications
