+++ registration/applications.py
@@ -6,7 +6,15 @@ from pydantic import validate_model
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.future import select
 
-from common.database import Application, ApplicationRead, ApplicationUpdate, with_db
+from common.authentication import with_user_id
+from common.database import (
+    Application,
+    ApplicationCreate,
+    ApplicationRead,
+    ApplicationUpdate,
+    with_db,
+)
+from common.kv import NamespacedClient, with_kv
 
 router = APIRouter()
 
@@ -24,8 +32,74 @@ async def list(
     return applications
 
 
+@router.post(
+    "/",
+    response_model=ApplicationRead,
+    status_code=HTTPStatus.CREATED,
+    name="Create application",
+)
+async def create_application(
+    values: ApplicationCreate,
+    id: str = Depends(with_user_id),
+    db: AsyncSession = Depends(with_db),
+    kv: NamespacedClient = Depends(with_kv("autosave")),
+):
+    """
+    Create a new application attached to the currently authenticated participant
+    """
+    application = Application.from_orm(values, {"participant_id": id})
+    db.add(application)
+    await db.commit()
+
+    # Delete the auto-save data
+    await kv.delete(str(id))
+
+    return application
+
+
+@router.get(
+    "/autosave",
+    response_model=ApplicationUpdate,
+    name="Get an in-progress application",
+)
+async def get_autosave_application(
+    id: str = Depends(with_user_id), kv: NamespacedClient = Depends(with_kv("autosave"))
+):
+    """
+    Get the data for an in-progress application
+    """
+    autosave = await kv.get(str(id), is_json=True)
+    if autosave:
+        return autosave
+
+    raise HTTPException(status_code=HTTPStatus.NOT_FOUND, detail="not found")
+
+
+@router.put(
+    "/autosave",
+    status_code=HTTPStatus.NO_CONTENT,
+    name="Save an in-progress application",
+)
+async def autosave_application(
+    values: ApplicationUpdate,
+    id: str = Depends(with_user_id),
+    db: AsyncSession = Depends(with_db),
+    kv: NamespacedClient = Depends(with_kv("autosave")),
+):
+    """
+    Save an in-progress application
+    """
+    # Prevent auto-saving if already applied
+    application = await db.get(Application, id)
+    if not application:
+        await kv.set(
+            str(id),
+            values.dict(exclude_unset=True, exclude_none=True, exclude_defaults=True),
+        )
+
+
 @router.get("/{id}", response_model=ApplicationRead, name="Read application")
-async def read(id: int, db: AsyncSession = Depends(with_db)):
+async def read(id: str, db: AsyncSession = Depends(with_db)):
     """
     Returns a single application by id
     """
@@ -36,7 +110,7 @@ async def read(id: int, db: AsyncSession = Depends(with_db)):
 
 
 @router.put("/{id}", response_model=ApplicationRead, name="Update application")
-async def update(id: int, info: ApplicationUpdate, db: AsyncSession = Depends(with_db)):
+async def update(id: str, info: ApplicationUpdate, db: AsyncSession = Depends(with_db)):
     """
     Updates an application by id
     """
@@ -59,7 +133,7 @@ async def update(id: int, info: ApplicationUpdate, db: AsyncSession = Depends(wi
 
 
 @router.delete("/{id}", status_code=HTTPStatus.NO_CONTENT, name="Delete application")
-async def delete(id: int, db: AsyncSession = Depends(with_db)) -> None:
+async def delete(id: str, db: AsyncSession = Depends(with_db)) -> None:
     """
     Deletes an application by id
     """
